+++ code/spec/spec_objs_weapons.cc
@@ -1507,7 +1507,7 @@ int brokenBottle(TBeing *vict, cmdTypeT cmd, const char *arg, TObj *o, TObj *)
 int boneStaff(TBeing *vict, cmdTypeT cmd, const char *arg, TObj *o, TObj *)
 {
   TBeing *ch;
-  wearSlotT part_hit = wearSlotT((int) arg);
+  wearSlotT part_hit = wearSlotT((long) arg);
 
   ch = genericWeaponProcCheck(vict, cmd, o, 0);
   if (!ch)
@@ -2271,7 +2271,7 @@ int gnomeTenderizer(TBeing *vict, cmdTypeT cmd, const char *arg, TObj *o, TObj *
     if (!vict || !(vict == ch->fight()))
       return FALSE;
     
-    wearSlotT part = wearSlotT(int(arg));
+    wearSlotT part = wearSlotT(long(arg));
     
     act("A bar from $p swings around and meets $N's flesh flush at its slightly-hollowed end.",
         false, ch, o, vict, TO_CHAR, ANSI_CYAN);
